+++ src/dr/app/gui/chart/ScatterPlot.java
@@ -30,8 +30,8 @@ import dr.stats.Variate;
 import java.awt.*;
 import java.awt.geom.Point2D;
 import java.awt.geom.Rectangle2D;
-import java.util.Vector;
 import java.util.Set;
+import java.util.Vector;
 
 /**
  * Description:	A scatter plot.
@@ -63,7 +63,7 @@ public class ScatterPlot extends Plot.AbstractPlot {
     /**
      * Constructor
      */
-    public ScatterPlot(double[] xData, double[] yData) {
+    public ScatterPlot(Double[] xData, Double[] yData) {
         super(xData, yData);
         setMarkStyle(CIRCLE_MARK, 5, new BasicStroke(1),
                 Color.black, Color.yellow);
@@ -152,8 +152,8 @@ public class ScatterPlot extends Plot.AbstractPlot {
 
         int n = xData.getCount();
         for (int i = 0; i < n; i++) {
-            x = (float) transformX(xData.get(i));
-            y = (float) transformY(yData.get(i));
+            x = (float) transformX((Double) xData.get(i));
+            y = (float) transformY((Double) yData.get(i));
 
             if (selectedPoints.contains(i)) {
                 drawMarkHilighted(g2, x, y);
